Rename genelpa to module_genelpa
